requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# LLM Client (OpenAI-compatible)
openai==1.51.0
httpx==0.27.2

# Browser Automation
playwright==1.48.0

# Document Parsing
PyMuPDF==1.24.11
python-docx==1.1.2
openpyxl==3.1.5

# Task Scheduling
apscheduler==3.10.4

# Email
aiosmtplib==3.0.2

# Security
cryptography==43.0.1

# Environment
python-dotenv==1.0.1
